refactor: split proposalDetail into Card components
