+++ frontend/svelte/src/App.svelte
@@ -11,7 +11,7 @@
   import { onDestroy } from "svelte";
   import { AuthStore, authStore } from "./lib/stores/auth.store";
   import Wallet from "./routes/Wallet.svelte";
-  import ProposalDetails from "./routes/ProposalDetails.svelte";
+  import ProposalDetails from "./routes/ProposalDetail.svelte";
   import { routeStore } from "./lib/stores/route.store";
   import { AppPath } from "./lib/constants/routes.constants";
   import Toasts from "./lib/components/ui/Toasts.svelte";
