+++ frontend/svelte/src/lib/constants/routes.constants.ts
@@ -5,5 +5,5 @@ export enum AppPath {
   Proposals = "/#/proposals",
   Canisters = "/#/canisters",
   Wallet = "/#/wallet",
-  ProposalDetails = "/#/proposal",
+  ProposalDetail = "/#/proposal",
 }
